+++ map_fold/src/edu/ucsd/map_fold/common/WorkerClient.java
@@ -18,7 +18,7 @@ public class WorkerClient extends GenericClient<WorkerInterface> implements Work
     public void sendToken(int target, int tokenId, int version) throws RemoteException {
         call( w -> {try{w.sendToken(target, tokenId, version); return null;}catch (RemoteException e){ return e; }} );
     }
-    public void loadData(String filePath, int offset, int count) throws RemoteException {
+    public void loadData(String filePath, Long offset, Long count) throws RemoteException {
         call( w -> {try{w.loadData(filePath, offset, count); return null;}catch (RemoteException e){ return e; }} );
     }
